included web3.min.js version 1.2.4
Since people would be git cloning this app, I thought it would be better if the web3.min.js file is also bundled together. Since it would be in the same folder as the index.html file, the script src would work without the user downloading and putting the file in the folder.